Optionally limit parallelization
While it works fine to make thousands of connections in parallel, users
may not want to run all commands in parallel. Give them the option to
limit parallelization.

This required a rewrite of the ui/printing functions to keep the
progress message at the bottom instead of mixing debug and progress
output.