Turn the execute loop into a two-phase loop: parse and execute
Instead of running things as they come in, we now first parse all the
arguments into a set of commands, then we run them in a loop. This is in
preparation for interactive and scripted mode, which will parse things
in a different way, but we want to have consistency when running.