Convert to a cobra app
Wanting to do the config with viper, and given the weirdness around
modes and config, it actually makes more sense to convert this to a
cobra app and simplify things.

This does change command line syntax, for the better :) Also a bunch of
small bugs popped up while validating this change, such as misuse of
loglevels.