Run ssh.Dial in a goroutine and manually time it out to workaround a go bug
Sometimes net.Dial will return a connection well before a connection has
been established, or even can be established. In these cases, the
client/server handshake deep inside net.Dial will hang forever, as it's
not governed by the timeout.

So let's run ssh.Dial in a goroutine and manually work on timeouts.
This also removes the need for the timeout2/not finishing properly
kludge.