Introduce the scripting engine and get rid of the global UI
The scripting engine ties the runner and UI together and executes
user-provided commands on either of them. This is necessarily a large
commit, as all commands and they way the parts talk to each other
change.

Moreover, we break the ability to set output filters. This will be
restored shortly.