Add an AWS provider and make it possible to not compile in certain providers
The aws provider supports multiple accounts, partitions and regions. It
will auto-detect the regions to use and by default query them all. This
can be overridden in the config. It will also magically load if
credentials are detected in the environment.

As it makes the binary significantly larger, you can opt out of building
it with the no_aws compile tag. The same now also applies for the consul
provider, and the autoloading mechanism is also used to make the putty
provider go away on non-windos os'es.