No global state for the ssh agent connection
In the next commit we'll have to wrap the SSH agent to do more efficient
signing. For this to be simple, we remove some technical debt here
around working with the agent. As a bonus, there's now a warning when
there's ssh agent trouble instead of each host erroring out.