Optionally wrap the ssh client object for faster authentication cycles
The built-in agent client waits for a response before sending the next
request. We don't always need to do that, ssh-agent will send responses
in-order, so we just need to make sure we return the responses to the
right requester and we can spam the agent as fast as possible.