test/sharness: normalize output more
In 7dfc9fe (test/sharness: normalize -x output before testing
it, 2019-06-27), we tried to normalize -x output before we
would test it, but it looks like this is sometimes not enough.
So let's also normalize the '+' characters and possibly the
following space characters that -x can output.

This output is influenced by the $PS4 variable, so a better
solution might be to change it, but for now let's try not to
touch it.

Fixes: https://github.com/chriscool/sharness/issues/104