workflows/ci: install IO::Pty on Ubuntu
IO::Pty is needed to perform tests that use test-terminal.

We do not install it on Windows, as IO::Tty mentions on its
homepage that Windows is only supported under the Cygwin
environment (see https://metacpan.org/pod/IO::Tty), and
GitHub actions use the git-bash shell, which is
MSYS/MinGW based.

We do not install it on MacOS, as according to
https://github.com/git/git/commit/339e5638b0a test-terminal
gets stuck reading from the pty master every once in a while,
and we don't want to risk test timing out.

Also if it works on other platforms it should work on Windows
and MacOS too, if it's properly installed and if there is no
weird behavior on the platform.