Add vim syntax file
Also, add a simple rule to set *.t files as sharness.

Do not check the filetype initially, so we can do 'set ft=sh.sharness'.
This will first load all the configurations for sh, then for sharness,
that we can preserve the specific indentation configured for sh, and
override it for sharness.

I have tested this syntax in many test files and works fine,
unfortunately not in test/sharness.t.

Signed-off-by: Felipe Contreras <felipe.contreras@gmail.com>