The cache provider moves to its own package
Unlike the other providers, we expect this to be used by other
providers, so we keep NewFromProvider pubic (and use it in the aws and
consul providers).

This did require some rejigging in registry.go, and herd.Cache now
becomes an interface, opening the door for alternative cache
implementations as well.

Lastly, this fixes a bug that caused data to be cached despite errors
being returned by the source providers.