Get rid of CacheMessage and its channel
It was never about caches, it was about loading data. For this we now
pass a callback. This callback may always be called when done loading
and the error that is passed will be shown to the user. It may be
called when starting to load data, but that should only be done when
expecting the load not to be near-instant (e.g. when calling the aws
api).

Lastly, we also redo the output to not talk about cache
messages and stop showing the loading messages when we're not connected
to a tty.