Add a plugin system based on hashicorp/plugin
This lets you run host providers outside the main process with no
changes to the provider code and 10 lines of code to run the provider as
a standalone program.

An example provider that can be run inside and outside the main process,
and the 10 lines to turn this provider into a standalone executable,
have been provided.