When we have host keys for a host, set the allowed host key algorithms
This helps e.g. putty users, whode idea of best algorithm is different
from openssh/golang. When a user only has an ed25519 key in their cached
known hosts and we try to use the defaults, go picks ecdsa and we don't
find a match unless we tell go only to use algorithms for which we have
a key.