Move the ssh agent handling out of the runner
Adding tokens in the agent path was the trigger to change this, as that
would create really weird interactions between runner and host. The
agent handling code moves to a separate package, setupScriptEngine (or
any custom client) becomes responsible for initializing the agent and
the host object knows how to expand tokens in the agent socket path.

The runner no longer does anything with the agent except passing it on
to the host objects.