Proper timeout for loading data from providers
Instead of a few providers having a configurable timeout, we now have
one timeout flag and use a context for timing out.

Unfortunately this requires a mildly gross hack in the transip provider,
as it doesn't natively support context cancellation. A slightly annoying
hack in the consul provider as one function doesn't take a context.  And
a really gross hack in the google provider as its context support is
broken and can't easily be monkeypatched in.