Give the provider plugin some much needed love
First the naming. Clearly I was learning go-plugin when I first wrote
this, as I used the name provider in places where plugin makes more
sense.

Then the interface. The way logging was initialized was another symptom
of just learning it. It's now initialized with a separate message and
much earlier. The server also forwards logs to the client as soon as
it's configured.

Then functionality. The provider plugin now implements the cache and
dataloader interfaces and forwards them to the actual provider if it
does so too.