Find ssh keys with a comment
Herd assumed that an ssh key's comment in the agent is the filename.
This is true if the ssh key does not have a comment, but false if the
ssh key does have one.

So now we still try to find the key by checking if the key has the
filename as comment, but if we find nothing this way, we read the public
key and match keys by content in the agent.