Make herd themeable
This started as just wanting to give hosts that ran the command but
exited with non-zero status, and hosts that never ran the command
because the timeout expired, a different color than the hosts where
connecting failed with an error.

But rather than adding yet another hardcoded color, the colors can now
be set in the config. A default theme mathing the existing output is
included.