Support connecting to hosts that stopped supporting ssh-rsa signature algorithms
Newer openssh supports ssh keys for authentication without supporting
the ssh-rsa signature algorithm. Instead, they support ssh-rsa-sha256 or
ssh-rsa-sha512.

The standard go agent client learned how to deal with this, but our pool
wrapper did not know about this yet. So we update our pool wrapper to
learn about this, borrowing code from x/crypto/ssh.

And to properly support this, we now need to parallelize SignWithFlags,
so we do.